git comments addressed
